+++ src/routes/[[lang]]/(protected)/verification/select-credential/+page.svelte
@@ -1,11 +1,7 @@
 <script lang="ts">
 	import HeaderWithBackButton from '$lib/components/molecules/HeaderWithBackButton.svelte';
 	import { m } from '$lib/i18n';
-	import { decodeSdJwt } from '$lib/openId4vci';
-	import { flip } from 'svelte/animate';
-	import { sineInOut } from 'svelte/easing';
 	import { slide } from 'svelte/transition';
-	import type { Credential } from '$lib/preferences/credentials';
 	import type { Feedback } from '$lib/utils/types.js';
 	import { verificationStore } from '$lib/verificationStore.js';
 	import dayjs from 'dayjs';
@@ -13,35 +9,15 @@
 	import { addVerificationActivity } from '$lib/preferences/activity.js';
 	import { verifyCredential } from '$lib/components/organisms/scanner/tools.js';
 	import { negativeFeedback } from '$lib/utils/index.js';
-	import { verificationResultsStore } from '$lib/verificationResultsStore.js';
+	import { verificationResultsStore, type VerificationResponse } from '$lib/verificationResultsStore.js';
 	import { goto } from '$app/navigation';
 	import DebugPopup from '$lib/components/organisms/debug/DebugPopup.svelte';
-	import {
-		debugDismiss,
-		debugPopup,
-		debugPopupContent
-	} from '$lib/components/organisms/debug/debug';
-	import { onMount } from 'svelte';
+	import { debugDismiss } from '$lib/components/organisms/debug/debug';
 
-	type VerificationResponse = {
+	type Verification = {
 		result: {
 			result: {
-				result: {
-					body: {
-						message: string;
-						registrationToken: string;
-					};
-					server_response: {
-						result: {
-							error: {
-								message: string;
-								code: string;
-							};
-						};
-						status: string;
-					};
-					url: string;
-				};
+				result: VerificationResponse
 				status: string;
 			};
 		};
@@ -51,82 +27,78 @@
 	export let data;
 	const { credentials } = data;
 
-	let selectedCredential: string | undefined =
-		credentials.length === 1 ? credentials[0].sdJwt : undefined;
+	let selectedCredential: number;
 
-	let verificationResponse: VerificationResponse;
+	let verification: Verification;
 	let scrollBox: HTMLDivElement;
 
-	const { info, post_without_vp } = $verificationStore;
+	const { vps, post_url } = $verificationStore;
 	const verificationFailed = 'verification failed';
 
-	const selectCredential = (credential: string | undefined) => {
+	const selectCredential = (credential: number) => {
 		selectedCredential = credential;
-		getSortedCredentials();
-		scrollBox.scrollIntoView({
-			behavior: 'smooth',
-			block: 'start'
-		});
+		// getSortedCredentials();
+		// scrollBox.scrollIntoView({
+		// 	behavior: 'smooth',
+		// 	block: 'start'
+		// });
 	};
 
 	const verify = async () => {
 		try {
-			verificationResponse = (await verifyCredential({
-				...post_without_vp,
-				body: {
-					...post_without_vp.body,
-					vp: selectedCredential!
-				}
-			})) as VerificationResponse;
-			const success = verificationResponse.result.result.result.server_response.status === '200';
+			verification = (await verifyCredential({
+				url: post_url,
+				body: vps[selectedCredential].presentation
+			})) as Verification;
+			
+			const responseSuccess = verification.result?.result?.status === '200';
+			const success = verification.result?.result?.result?.output?.[0] === 'OK';
 			await debugDismiss();
 			const date = dayjs().toString();
 			let feedback: Feedback = {};
-			if (!success) {
+			if (!responseSuccess || !success) {
 				feedback = negativeFeedback(
 					verificationFailed,
 					JSON.stringify({
 						serverResponse:
-							verificationResponse.result.result.result.server_response.result.error.code,
-						message: verificationResponse.result.result.result.server_response.result.error.message,
-						logs: verificationResponse.logs
-					})
+							verification.result?.result?.result,
+						logs: verification.logs
+					}, null, 2)
 				);
 			}
 			verificationResultsStore.set({
 				feedback,
 				date,
-				id: post_without_vp.body.id,
-				success
+				id: verification.result?.result?.result?.complete_transaction_id || '',
+				success: responseSuccess && success
 			});
-			await addVerificationActivity(post_without_vp.body.id, info, success);
-			log(JSON.stringify(verificationResponse));
+			log(JSON.stringify(verification));
 			return await goto('/verification/results');
 		} catch (e) {
 			verificationResultsStore.set({
 				feedback: negativeFeedback(verificationFailed, JSON.stringify(e)),
 				date: dayjs().toString(),
-				id: post_without_vp.body.id,
+				id: "none",
 				success: false
 			});
 			log(JSON.stringify(e));
-			await addVerificationActivity(post_without_vp.body.id, info, false);
+			// await addVerificationActivity(post_without_vp.body.id, info, false);
 			return await goto('/verification/results');
 		}
 	};
 
-	let sortedCredentials: Credential[];
-	const getSortedCredentials = () => {
-		if (selectedCredential) {
-			sortedCredentials = [
-				credentials.find((c) => c.sdJwt === selectedCredential),
-				...credentials.filter((c) => c.sdJwt !== selectedCredential)
-			] as Credential[];
-			return 
-		}
-		sortedCredentials = credentials;
-	};
-	onMount(getSortedCredentials);
+	// let sortedCredentials: Credential[];
+	// const getSortedCredentials = () => {
+	// 	if (selectedCredential) {
+	// 		sortedCredentials = [
+	// 			credentials.find((c) => c.sdJwt === selectedCredential),
+	// 			...credentials.filter((c) => c.sdJwt !== selectedCredential)
+	// 		] as Credential[];
+	// 		return;
+	// 	}
+	// 	sortedCredentials = credentials;
+	// };
+	// onMount(getSortedCredentials);
 </script>
 
 <HeaderWithBackButton>
@@ -141,39 +113,33 @@
 				description={m.novel_elegant_capybara_twist({ length: credentials.length })}
 			/>
 			<d-vertical-stack>
-				{#if sortedCredentials}
-					{#each sortedCredentials as credential, index (credential.sdJwt)}
-						<d-verification-card
-							class:opacity-60={selectedCredential && selectedCredential !== credential.sdJwt}
-							class="transition-opacity duration-500"
-							selected={selectedCredential === credential.sdJwt}
-							relying-party={credential.issuerUrl}
-							verifier={credential.issuer}
-							logo={credential.logo.uri}
-							flow={credential.display_name}
-							on:click={() => selectCredential(credential.sdJwt)}
-							aria-hidden
-							animate:flip={{ duration: 400, easing: sineInOut }}
-						>
-							{#await decodeSdJwt(credential.sdJwt) then sdJwt}
+				{#each credentials as credential, index}
+					<d-verification-card
+						class:opacity-60={selectedCredential && selectedCredential !== index}
+						class="transition-opacity duration-500"
+						selected={selectedCredential === index}
+						relying-party={credential.issuer}
+						verifier={credential.issuer}
+						flow={credential.type}
+						on:click={() => selectCredential(index)}
+						aria-hidden
+					>
+						<!-- {#await decodeSdJwt(credential.sdJwt) then sdJwt}
 								{#each sdJwt.credential.disclosures as disclosure}
 									<d-definition title={disclosure[1]} definition={disclosure[2]} dotted
 									></d-definition>
 								{/each}
-							{/await}
-						</d-verification-card>
-					{/each}
-				{/if}
+							{/await} -->
+					</d-verification-card>
+				{/each}
 				<div class="pb-56" />
 			</d-vertical-stack>
 		</d-vertical-stack>
 	</div>
-	{#if selectedCredential}
+	{#if selectedCredential !== undefined}
 		<div class="ion-padding fixed bottom-0 h-40 w-full bg-surface" transition:slide>
 			<d-vertical-stack>
-				<d-button on:click={verify} aria-hidden expand color="accent" disabled={!selectedCredential}
-					>{m.Verify()}</d-button
-				>
+				<d-button on:click={verify} aria-hidden expand color="accent">{m.Verify()}</d-button>
 				<d-button expand aria-hidden>{m.Decline()}</d-button>
 			</d-vertical-stack>
 		</div>