+++ src/routes/[[lang]]/(protected)/wallet/+page.svelte
@@ -23,7 +23,7 @@
 			<d-illustration illustration="empty-wallet" />
 		</d-empty-state>
 	{:else}
-		<div class="grid md:grid-cols-2 grid-cols-1 gap-12">
+		<div class="grid grid-cols-1 gap-12 md:grid-cols-2">
 			{#each credentials as credential}
 				{@const expirationDate = dayjs.unix(credential.expirationDate).format('DD.MM.YYYY HH:mm')}
 				<button on:click={() => goto(`/${credential.id}/credential-detail`)} class="relative">
@@ -45,11 +45,18 @@
 						issuedByLabel={m.Issued_by()}
 						expirationLabel={'Exp'}
 						class="max-w-44"
-						>{#await decodeSdJwt(credential.sdJwt) then sdjwt}
-							{#each sdjwt.credential.disclosures as disclosure}
-								<d-badge>{disclosure[1]}</d-badge>
+					>
+						{#if credential.type === 'ldp_vc'}
+							{#each Array.from(Object.entries(credential.ldpVc.credentialSubject)) as disclosure}
+								<d-badge>{disclosure[0]}</d-badge>
 							{/each}
-						{/await}
+						{:else if credential.type === 'sdjwt'}
+							{#await decodeSdJwt(credential.sdJwt) then sdjwt}
+								{#each sdjwt.credential.disclosures as disclosure}
+									<d-badge>{disclosure[1]}</d-badge>
+								{/each}
+							{/await}
+						{/if}
 					</d-credential-card>
 				</button>
 			{/each}