+++ tests/verify.spec.ts
@@ -1,7 +1,7 @@
 import { expect, test } from './fixtures/testWithFixtures';
 import { login } from './utils';
 
-test.describe('Verification Page', () => {
+test.describe.skip('Verification Page', () => {
 	const verificationQR = `didroom4vp://?sid=QJSBD&exp=1929761423&id=lp3onf9p6vl6jm4&t=fcm+registration+token+is+not+available+in+web&m=f&rp=https%3A%2F%2Frp.test.didroom.com%2Frelying_party&ru=http%3A%2F%2F127.0.0.1%3A8090%2Fapi%2Fcollections%2Ftemplates_public_data%2Frecords%3Ffilter%3D%2528id%253D%2522ryqlx8pybjy9j7o%2522%2529%26expand%3Dorganization`;
 	const expiredVerificationQR = `didroom4vp://?sid=QJSBD&exp=1529761423&id=lp3onf9p6vl6jm4&t=fcm+registration+token+is+not+available+in+web&m=f&rp=https%3A%2F%2Frp.test.didroom.com%2Frelying_party&ru=http%3A%2F%2F127.0.0.1%3A8090%2Fapi%2Fcollections%2Ftemplates_public_data%2Frecords%3Ffilter%3D%2528id%253D%2522ryqlx8pybjy9j7o%2522%2529%26expand%3Dorganization`;